pyproject.toml
[project]
name = "spea-lab-iot"
version = "0.1.0"
description = "SPEA Lab IoT project"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=46.0.5",
    "paho-mqtt>=2.0",
    "pycryptodome>=3.23.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/spea_lab_iot", "src/mqtt_demo"]

[tool.ruff]
line-length = 88
target-version = "py311"

[dependency-groups]
dev = []
